Split analytics-ipc.ts into analytics-ipc-channel-name.ts and google-analytics-4-ipc.ts
