+++ src-main/electron-main.ts
@@ -4,7 +4,7 @@ import * as path from 'path';
 import * as url from 'url';
 import * as electronWebPreferences from '../electron-util/electron-web-preferences';
 import '../electron-util/configure-electron-remote-in-main-process';
-import { GoogleAnalytics4IpcMain } from '../src-shared/analytics/analytics-ipc';
+import { GoogleAnalytics4IpcMain } from '../src-shared/analytics/ipc/google-analytics-4-ipc';
 import { UniversalAnalyticsWrapper } from '../src-shared/analytics/universal-analytics-wrapper';
 import { Logger } from '../src-shared/log/logger';
 import { LogFileConfig } from '../src-shared/log/log-file-config';
