+++ src-shared/analytics/ipc/analytics-ipc-channel-name.ts
@@ -1,0 +1,4 @@
+export class AnalyticsIpcChannelName {
+  public static readonly googleAnalytics4TrackEvent = 'google-analytics-4-track-event';
+  public static readonly universalAnalyticsTrackEvent = 'universal-analytics-track-event';
+}
